Adding indexes and preserving precision

+++ src/main/java/uk/dansiviter/scd/entity/WindowEntity.java
@@ -1,14 +1,15 @@
 package uk.dansiviter.scd.entity;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.time.OffsetDateTime;
 
 public class WindowEntity {
 	private final Instant start;
 	private final Instant end;
-	private final Long value;
+	private final BigDecimal value;
 
-	public WindowEntity(OffsetDateTime start, OffsetDateTime end, Long value) {
+	public WindowEntity(OffsetDateTime start, OffsetDateTime end, BigDecimal value) {
 		this.start = start.toInstant();
 		this.end = end.toInstant();
 		this.value = value;
@@ -22,7 +23,7 @@ public class WindowEntity {
 		return end;
 	}
 
-	public Long getValue() {
+	public BigDecimal getValue() {
 		return value;
 	}
 }
